requirements.txt
tensorflow==4.8.0
keras==2.13.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
pillow==10.0.0
opencv-python==4.8.0.76
speech-recognition==3.10.0
pyttsx3==2.90
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
paho-mqtt==1.6.1
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
redis==5.0.0
python-socketio==5.9.0
pytest==7.4.2
pytest-cov==4.1.0
joblib==1.3.2
xgboost==2.0.0
